Implement play/pause, next, and previous commands.
This fixes #3. It also causes #13. The go_to_prev function of playlist
is called twice, which causes the bug. This is because when the stream
is started, the playlist is told to go to the next song, so it needs to
be on the one before the one that is going to play.