Decrease CPU usage of RFSMP.
Added 20ms sleep to decrease the CPU usage. Also, made the UI only
render when the song location or song duration have changed. Although
this fixes #11, it raises a problem with #5. This code relies on the
fact that the times only change once every second, or when the song
changes. If the times become more accurate, then the UI would go back to
rendering a lot, increasing CPU usage.